empty editor.js, prepare for deletion
