+++ h5p/development/H5PEditor.FormulaAppletEditor-0.12/library.json
@@ -2,7 +2,7 @@
   "title": "H5PEditor.FormulaAppletEditor",
   "majorVersion": 0,
   "minorVersion": 12,
-  "patchVersion": 37,
+  "patchVersion": 47,
   "runnable": 0,
   "machineName": "H5PEditor.FormulaAppletEditor",
   "author": "Rudolf Grossmann",
