+++ main/nextpatchnumber.txt
@@ -1,1 +1,1 @@
-38
+48
