+++ main/runcopy.ps1
@@ -1,2 +1,3 @@
-npm run build
-.\copyBundles.cmd
+.\patch.sh
+npm run h5p
+# .\copyBundles.cmd
