+++ main/src/js/config.json.ori
@@ -5,5 +5,5 @@
     "unit_replacement": "\\textcolor{blue}{",
 	"patchversion": "__PATCHVERSION__",
     "debug": "false",
-    "htmloutput": "true"
+    "htmloutput": "false"
 }
